requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
loguru==0.7.0
psycopg2-binary==2.9.6
